requirements.txt
pandas>=2.0
numpy>=1.24
pyarrow>=12.0
fastapi>=0.110
uvicorn>=0.29
kafka-python>=2.0
--extra-index-url https://download.pytorch.org/whl/cpu
torch
